+++ lib/src/enums/log_type.dart
@@ -11,4 +11,5 @@ enum LogType {
   debug,
   mvvm,
   bloc,
+  test,
 }